ajustes de incompatibilidades pos integraÃcao

+++ src/main/java/faculdade/mercadopago/usecase/IPagamentoUseCase.java
@@ -5,12 +5,10 @@ import faculdade.mercadopago.entity.Pedido;
 import faculdade.mercadopago.entity.pagamento.ConfirmacaoPagamentoRes;
 import faculdade.mercadopago.entity.pagamento.QrCodeRes;
 
-import java.math.BigDecimal;
-
 public interface IPagamentoUseCase {
     QrCodeRes processarQrCode(QrCodeRequest request);
 
-    void salvarPagamento(Pedido pedido, BigDecimal valor);
+    void salvarPagamento(Pedido pedido, Double valor);
 
 
     ConfirmacaoPagamentoRes consultarPagamento(String id);
